Optimize Ball behaviour
Using OnCollisionStay() is really bad idea as it gets called on every frame when it's touching another object, combined with known performance issues with Udon in general.

OnCollisionStay() can often be replaced by caching a boolean from OnCollision<Enter/Exit>()

FixedUpdate() can get called up to 4 times / frame and can become extremely performance heavy really quickly.

Accessing objects outside of the local UdonBehaviour is a "blacklisted" call, meaning that UdonVM internally ensures that you are not trying to access the game's internal objects, which results in a significant overhead, this is really bad when combined with high-frequency methods (e.g. OnTriggerStay() and FixedUpdate())